backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
fpdf2==2.8.3
